+++ src/app/docs/layout/demo-divider/demo-divider.component.ts
@@ -41,7 +41,18 @@ export class DemoDividerModule {
                     },
                 }
             ],
-            apis: [],
+            api_groups: [
+                {
+                    name: '',
+                    apis: [
+                        {
+                            label: '',
+                            type: '',
+                            description: '',
+                        }
+                    ],
+                }
+            ],
         };
     }