+++ src/app/guides/get-started/get-started.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { DemoPageModel } from '../../shared/model';
 
 @Component({
     selector: 'app-get-started',
@@ -27,18 +26,22 @@ ng add ngx-slice-kit`;
 })
 export class SharedModule {
 }`;
-    batchModuleExample = `import { ButtonsModule, ModalsModule, DropdownsModule } from 'ngx-slice-kit';
+    batchModuleExample = `import { ThemeModule, ButtonsModule, ModalsModule, DropdownsModule } from 'ngx-slice-kit';
 
 @NgModule({
     imports: [
         CommonModule,
         FormsModule,
-        // import slice-kit by feature modules
+        // theme module is required
+        ThemeModule,
+        // import feature modules
         ButtonsModule,
         ModalsModule,
         DropdownsModule,
     ],
+    // optional
     exports: [
+        ThemeModule,
         ButtonsModule,
         ModalsModule,
         DropdownsModule,
@@ -47,6 +50,7 @@ export class SharedModule {
 export class SharedModule {
 }`;
     componentModuleExample = `import {
+    ThemeModule,
     ToggleModule,
     ButtonModule,
     AlertModule,
@@ -56,6 +60,9 @@ export class SharedModule {
 } from 'ngx-slice-kit';
 
 const sliceKitImports = [
+    // theme module is required
+    ThemeModule,
+    // import component modules
     ToggleModule,
     ButtonModule,
     AlertModule,
@@ -71,6 +78,7 @@ const sliceKitImports = [
         // add slice-kit imports
         ...sliceKitImports
     ],
+    // optional
     exports: [
         // share components through module
         ...sliceKitImports
@@ -85,6 +93,14 @@ export class SharedModule {
     <router-outlet></router-outlet>
 </div>
 `;
+    stylesExample = `@import './libs/ngx-slice-kit/src/lib/core/styles/core';`;
+    angularJson = `                        // ...
+// Double check that your core application styles file used in
+// "ng build" and/or "ng serve" commands
+                        "styles": [
+                            "src/styles.scss"
+                        ],
+                        // ...`;
 
     constructor() {
     }